backend/requirements.txt
# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# LangChain
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# FastAPI Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
httplib2>=0.22.0
